feat: Migrate ESLint configuration to a module
The changes made in this commit migrate the ESLint configuration from a
JSON file (.eslintrc.json) to a JavaScript module (eslint.config.js).
This allows for more flexibility and customization in the ESLint
configuration, such as the ability to use dynamic values or
programmatically generate the configuration.

The new configuration file includes the same rules as the previous
configuration, with the addition of the `languageOptions` object to
specify the ECMAScript version, source type, and global variables.